Move/absorb auxlib (#11025)
* Moved auxlib

* Update auxlib imports

* Fix auxlib flake8 failures

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Documenting why this move occurred
