+++ conda/core/solve.py
@@ -15,8 +15,8 @@ from .link import PrefixSetup, UnlinkLinkTransaction
 from .prefix_data import PrefixData
 from .subdir_data import SubdirData
 from .. import CondaError, __version__ as CONDA_VERSION
-from .._vendor.auxlib.decorators import memoizedproperty
-from .._vendor.auxlib.ish import dals
+from ..auxlib.decorators import memoizedproperty
+from ..auxlib.ish import dals
 from .._vendor.boltons.setutils import IndexedSet
 from .._vendor.toolz import concat, concatv, groupby
 from ..base.constants import DepsModifier, UNKNOWN_CHANNEL, UpdateModifier, REPODATA_FN
